move sys-info into vpnLib
