+++ nym-vpn-core/crates/nym-vpn-lib/src/platform/mod.rs
@@ -138,6 +138,8 @@ async fn configure_lib(
     sentry_monitoring: bool,
 ) -> Result<(), VpnError> {
     let network = environment::current_environment_details().await?;
+    let os = crate::SysInfo::new();
+    os.raw_display(true);
     if sentry_monitoring {
         let mut guard = SENTRY_CLIENT.lock().await;
         *guard = sentry_monitoring::init();
@@ -149,6 +151,8 @@ async fn init_logger(path: Option<PathBuf>, debug_level: Option<String>, sentry_
     let default_log_level = env::var("RUST_LOG").unwrap_or("info".to_string());
     let log_level = debug_level.unwrap_or(default_log_level);
     tracing::info!("Setting log level: {log_level}, path?: {path:?}");
+    let os = crate::SysInfo::new();
+    os.display(true);
     if sentry_monitoring {
         let mut guard = SENTRY_CLIENT.lock().await;
         *guard = sentry_monitoring::init();
